+++ NQueens/bin/Board.class
