+++ NQueens/bin/Frame.class
