+++ NQueens/bin/NQueens.class
