+++ NQueens/bin/Queen.class
