+++ NQueens/bin/StackSolver.class
