+++ NQueens/res/WQueen.png
